Nkandla

+++ src/main/java/com/bbd/nkandla/Amphitheatre.java
@@ -1,13 +1,13 @@
 package com.bbd.nkandla;
 
 public class Amphitheatre {
-	private double area;
+	private int area;
 	
 	public Amphitheatre (int area){
 		this.area= area;
 	}
 	
-	public double getArea(){
+	public int getArea(){
 		return this.area;
 	}
